Use COSE_Encrypt to signal key derivation for PSK
The KDF information is signaled in the recipient structure of
the COSE_Encrypt structure. For the PSK case, it is important
that RS knows which KDF algorithm to use, the nonce used by AS
to calculate the session key (and possibly the kid of the shared
secret for the HKDF).