Issue #10: Use kid by default for psk_identity
The psk_identity now may contain a (currently base64-encoded) key
identifier as specified in the kid parameter for a COSE_Key. If no
matching key was found by the receiving server, it may optionally
process the decoded contents of the psk_identity as if it was an
access token to be uploaded with the authorization information endpoint.

This change resolves issue #10 in favor of option b.2. Code-complexity
also decreases as the processing as access token now is optional (cf.
issue #14).