Issue #12: Added text on delayed session termination
When no more valid access tokens are available for a DTLS session
a resource server may want to cleanup state. To give the corresponding
peer the opportunity to update expired access tokens, it can delay
session termination and signal the key identifier to use for refresh.

This change addresses issue #12.