Clarify selection of RPK vs. PSK for profile coap_dtls
In a previous review, it was suggested to change the title and
introductory language to treat RPK and PSK as different profiles.
Comments from the ACE WG at IETF-100 suggested that coap_dtls should
be treated as a single profile. The title and introductory text
therefore have been changed back talk about one profile instead of
two.

This change reverts parts of commit 90b96b9.