Describe protection of Client-to-AS request in RPK mode
Added a sentence that forces the Client in RPK mode to use the same
key to protect the communication with the AS as is specified in the
cnf field of the Client-to-AS request. By mandating this, the client
cannot request an access token for someone else. The AS therefore
knows exactly which entity is being authorized.