Clarify protection of Client-AS communication
As required by Appendix C of draft-ietf-ace-oauth-authz-10 [1], the
profile needs to specify how the communication between the client
and the authorization server is protected. As the response may contain
a plaintext symmetric key to be used in the PSK mode, the requirement
here is that both parties must authenticate each other and the
communication must be confidential and integrity-protected. This could
be achieved by pre-configuring the client with AS's credentials and
doing an OAuth client registration.

[1] https://tools.ietf.org/html/draft-ietf-ace-oauth-authz-10#appendix-C