clarify CBOR serialization for HKDF input
It has been pointed out on the mailing list[1] that canonical CBOR
will be superseded in favor of preferred serialization and
deterministic encoding according to Section 4 of
draft-ietf-cbor-7049bis. This change updates the specification of the
CBOR data used as input for the key derivation function accordingly.
Moreover, the treatment of the access_token contents as opaque data
for the key derivation is clarified.

[1] See https://mailarchive.ietf.org/arch/msg/ace/KdnDNTaGe11SOnqGchwhy56lQuc/
    for the discussion thread on the ACE mailing list.