Add example for access token ordering
There is no natural strict order on the recency of access tokens. The
framework document lists a number of mechanisms the resource server
can use to guess if one access token is more recent than the
other. One global order could be defined using the cti claim which
is mentioned here as an example way forward.

This issue has been raised by Paul Kyzivat in the Gen-ART review of
2020-07-19.