Split structure of RPK mode and PSK mode sections
The structure of sections 3.2 and 3.3 has been split into two
subsections for each section:

The first describes the interaction between the client and the
authorization server.

The second describes the interaction between the client and the
resource server.

This split has been suggested by Paul Kyzivat in the Gen-ART review of
2020-07-19.